+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 * **NEW:** The configurable [request pipeline](README.md#request-pipeline) allows to limit the number of HTTP requests made in parallel.
 
+* Runtime parameter values are now displayed before running the steps of a scenario with the debug log level.
+  An [obfuscate parameter option](README.md#parameters-obfuscate) has been added for those parameters that you might not want to log in cleartext.
+  This option is turned on by default for parameters whose name contains the words `password`, `secret` or `authToken` (case-insensitive).
+
 * *BUGFIX:* the log level option could not be set in the scenario object because the command line program would always override it to the info level.
 
 * *BUGFIX:* the scenario listing would unnecessarily be displayed when using the run or info commands with an argument and a loading error occurred (e.g. due to an invalid scenario object).